+++ src/index.ts
@@ -15,5 +15,4 @@ export * from './plugins';
 //TODO: use inject in base classes also in plugins
 //TODO: maybe unify id and name and use only one (ID)??
 //TODO: maybe add ordering to permanent grid initializer
-//TODO: make providers functions for tokens
 //TODO: data loader for incremental data, maybe just options